requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
blobfile==2.0.1
breadability==0.1.20
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
dataclasses-json==0.5.7
docopt==0.6.2
ffmpeg-downloader==0.2.0
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.9.0
Flask==2.2.3
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.12.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.103
lxml==4.9.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
more-itertools==9.1.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.24.2
openai==0.27.0
openai-whisper @ git+https://github.com/openai/whisper.git@3e1780fd37686666f568be9c99f5b5e3e4f2eb92
packaging==23.0
pandas==1.5.3
pycountry==22.3.5
pycryptodomex==3.17
pydantic==1.10.5
python-dateutil==2.8.2
pytube==12.1.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
tenacity==8.2.2
tiktoken==0.3.0
tokenizers==0.13.2
torch==1.13.1
tqdm==4.64.1
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
